requirements.txt
numpy~=2.1.2
torchvision~=0.21.0+cu126
pillow~=11.0.0
torch~=2.6.0+cu126
pyiqa~=0.1.13
matplotlib~=3.10.1
tensorboardx~=2.6.2.2
pyyaml~=6.0.2
opencv-python~=4.11.0.86